builder-stake: move the instruction builders into the crate
The harness said to do this when something outside the tests needed them,
and something does. A caller in another repository has to post a bond to
test what it does with the account, and hand-encoding the accounts there
would put the processor's account order in two places, in two repositories,
with nothing to keep them in step.

A move, not a rewrite. The harness re-exports them, so every test that used
them reads unchanged, and all 21 still pass.
