+++ solana/programs/builder-stake/src/instruction/mod.rs
@@ -1,3 +1,5 @@
+pub mod builders;
+
 use borsh::{BorshDeserialize, BorshSerialize};
 use solana_pubkey::Pubkey;
