+++ solana/programs/builder-stake/tests/common/mod.rs
@@ -6,15 +6,14 @@
 #![allow(dead_code)]
 
 use doublezero_builder_stake::{
-    instruction::BuilderStakeInstructionData,
-    state::{self, BuilderStake, ProgramConfig},
+    state::{BuilderStake, ProgramConfig},
     DOUBLEZERO_MINT_KEY, ID,
 };
 use solana_loader_v3_interface::{get_program_data_address, state::UpgradeableLoaderState};
 use solana_program_test::{ProgramTest, ProgramTestContext};
 use solana_sdk::{
     account::Account,
-    instruction::{AccountMeta, Instruction, InstructionError},
+    instruction::{Instruction, InstructionError},
     program_pack::Pack,
     pubkey::Pubkey,
     signature::{Keypair, Signer},
@@ -281,157 +280,6 @@ impl TestSetup {
     }
 }
 
-//
-// Instruction builders.
-//
-
-fn encode(data: &BuilderStakeInstructionData) -> Vec<u8> {
-    borsh::to_vec(data).unwrap()
-}
-
-pub fn initialize_program(payer: &Pubkey) -> Instruction {
-    Instruction {
-        program_id: ID,
-        accounts: vec![
-            AccountMeta::new(*payer, true),
-            AccountMeta::new(ProgramConfig::find_address().0, false),
-            AccountMeta::new_readonly(system_program::ID, false),
-        ],
-        data: encode(&BuilderStakeInstructionData::InitializeProgram),
-    }
-}
-
-pub fn set_admin(admin: &Pubkey) -> Instruction {
-    Instruction {
-        program_id: ID,
-        accounts: vec![
-            AccountMeta::new_readonly(get_program_data_address(&ID), false),
-            AccountMeta::new_readonly(*admin, true),
-            AccountMeta::new(ProgramConfig::find_address().0, false),
-        ],
-        data: encode(&BuilderStakeInstructionData::SetAdmin(*admin)),
-    }
-}
-
-pub fn set_paused(admin: &Pubkey, paused: bool) -> Instruction {
-    use doublezero_builder_stake::instruction::{ProgramConfiguration, ProgramFlagConfiguration};
-
-    Instruction {
-        program_id: ID,
-        accounts: vec![
-            AccountMeta::new_readonly(*admin, true),
-            AccountMeta::new(ProgramConfig::find_address().0, false),
-        ],
-        data: encode(&BuilderStakeInstructionData::ConfigureProgram(
-            ProgramConfiguration::Flag(ProgramFlagConfiguration::IsPaused(paused)),
-        )),
-    }
-}
-
-pub fn set_tier_parameters(
-    admin: &Pubkey,
-    up_to_1gbps_2z_amount: u64,
-    up_to_5gbps_2z_amount: u64,
-    unmetered_2z_amount: u64,
-) -> Instruction {
-    use doublezero_builder_stake::instruction::ProgramConfiguration;
-
-    Instruction {
-        program_id: ID,
-        accounts: vec![
-            AccountMeta::new_readonly(*admin, true),
-            AccountMeta::new(ProgramConfig::find_address().0, false),
-        ],
-        data: encode(&BuilderStakeInstructionData::ConfigureProgram(
-            ProgramConfiguration::TierParameters {
-                up_to_1gbps_2z_amount,
-                up_to_5gbps_2z_amount,
-                unmetered_2z_amount,
-            },
-        )),
-    }
-}
-
-pub fn initialize_builder_stake(
-    builder: &Pubkey,
-    stake_index: u64,
-    committed_rate_bits_per_sec: u64,
-) -> Instruction {
-    let stake_key = BuilderStake::find_address(builder, stake_index).0;
-
-    Instruction {
-        program_id: ID,
-        accounts: vec![
-            AccountMeta::new_readonly(ProgramConfig::find_address().0, false),
-            AccountMeta::new(*builder, true),
-            AccountMeta::new(stake_key, false),
-            AccountMeta::new(state::find_2z_token_pda_address(&stake_key).0, false),
-            AccountMeta::new_readonly(DOUBLEZERO_MINT_KEY, false),
-            AccountMeta::new_readonly(spl_token_interface::ID, false),
-            AccountMeta::new_readonly(system_program::ID, false),
-        ],
-        data: encode(&BuilderStakeInstructionData::InitializeBuilderStake {
-            stake_index,
-            committed_rate_bits_per_sec,
-        }),
-    }
-}
-
-pub fn post_bond(
-    builder: &Pubkey,
-    stake_index: u64,
-    source_token_account: &Pubkey,
-    amount: u64,
-) -> Instruction {
-    let stake_key = BuilderStake::find_address(builder, stake_index).0;
-
-    Instruction {
-        program_id: ID,
-        accounts: vec![
-            AccountMeta::new_readonly(ProgramConfig::find_address().0, false),
-            AccountMeta::new_readonly(*builder, true),
-            AccountMeta::new(stake_key, false),
-            AccountMeta::new(state::find_2z_token_pda_address(&stake_key).0, false),
-            AccountMeta::new(*source_token_account, false),
-            AccountMeta::new_readonly(spl_token_interface::ID, false),
-        ],
-        data: encode(&BuilderStakeInstructionData::PostBond { amount }),
-    }
-}
-
-pub fn withdraw(
-    builder: &Pubkey,
-    stake_index: u64,
-    destination_token_account: &Pubkey,
-    amount: u64,
-) -> Instruction {
-    let stake_key = BuilderStake::find_address(builder, stake_index).0;
-
-    Instruction {
-        program_id: ID,
-        accounts: vec![
-            AccountMeta::new_readonly(ProgramConfig::find_address().0, false),
-            AccountMeta::new_readonly(*builder, true),
-            AccountMeta::new(stake_key, false),
-            AccountMeta::new(state::find_2z_token_pda_address(&stake_key).0, false),
-            AccountMeta::new(*destination_token_account, false),
-            AccountMeta::new_readonly(spl_token_interface::ID, false),
-        ],
-        data: encode(&BuilderStakeInstructionData::Withdraw { amount }),
-    }
-}
-
-/// Move a stake's hold expiry. Development builds only; the test binary is one.
-pub fn set_hold_expiry(admin: &Pubkey, builder: &Pubkey, stake_index: u64, at: i64) -> Instruction {
-    Instruction {
-        program_id: ID,
-        accounts: vec![
-            AccountMeta::new_readonly(ProgramConfig::find_address().0, false),
-            AccountMeta::new_readonly(*admin, true),
-            AccountMeta::new(BuilderStake::find_address(builder, stake_index).0, false),
-        ],
-        data: encode(&BuilderStakeInstructionData::SetHoldExpiry {
-            hold_expires_at: at,
-        }),
-    }
-}
+// The instruction builders moved into the crate, where a caller outside these tests can reach
+// them. Re-exported so the test files that use them read unchanged.
+pub use doublezero_builder_stake::instruction::builders::*;
